fix: fix chat_with_repo unit_test
